% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsrb_com.R
\name{zsrb_com}
\alias{zsrb_com}
\title{Calculate the Complex Standard Regression Based Change Score}
\usage{
zsrb_com(idf, t1, t2, group, ref = NULL, covs = NULL)
}
\arguments{
\item{idf}{Input data frame}

\item{t1}{Outcome variable scores at time point 1}

\item{t2}{Outcome variable scores at time point 2}

\item{group}{Group variable. If group membership changes over time, you can
add an array with groups, e.g., c("group_t1", "group_t2")}

\item{ref}{Reference group. If not set, the first factor level or lowest
value will be selected}

\item{covs}{One or more covariates that you want to take into account}
}
\description{
Calculate the Standard Regression Based Change Score as described by
Hammers, Suhrie, Dixon, Porter and Duff (2016).
https://doi.org/10.1093/arclin/acz076
The input data frame must be in wide format.
}
