% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsrb_sim.R
\name{zsrb_sim}
\alias{zsrb_sim}
\title{Calculate the Simple Standard Regression Based Change Score}
\usage{
zsrb_sim(idf, t1, t2, group, ref = NULL)
}
\arguments{
\item{idf}{Input data frame}

\item{t1}{Outcome variable scores at time point 1}

\item{t2}{Outcome variable scores at time point 2}

\item{group}{Group variable. If group membership changes over time, you can
add an array with groups, e.g., c("group_t1", "group_t2")}

\item{ref}{Reference group. If not set, the first factor level or lowest
value will be selected}
}
\description{
Calculate the Standard Regression Based Change Score as described by
Maassen, Bossema, and Brand (https://doi.org/10.1080/13803390802169059).
The input data frame must be in wide format.
}
