% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsrb_sim.R
\name{zsrb_sim}
\alias{zsrb_sim}
\title{Calculate the Simple Standard Regression Based Change Score}
\usage{
zsrb_sim(idf, t1, t2, group1, group2, ref, covs = NULL)
}
\arguments{
\item{idf}{Input data frame}

\item{t1}{Outcome variable scores at time point 1}

\item{t2}{Outcome variable scores at time point 2}

\item{group1}{Group variable at time point 1}

\item{group2}{Group variable at time point 2}

\item{ref}{Reference group}

\item{covs}{One or more covariates that you want to take into account}
}
\description{
Calculate the Standard Regression Based Change Score as described by
Maassen, Bossema, and Brand (https://doi.org/10.1080/13803390802169059).
The input data frame must be in wide format.
}
